fix: Move and rotation accelerations

+++ Engine/include/EngineScripting/commands/FollowCurve.h
@@ -28,9 +28,9 @@ private:
 	std::vector<glm::vec3> m_controlPoints; // Bezier curve control points 
 											 // (current position inserted in first index)
 
-	float m_timeStep;     // Steps to calculate the curve
-	float m_time;		  // Max time that should take to move
-	bool m_stopAtEnd;     // When arrive at location should stop?
+	int m_timeStep;     // # Steps to calculate the curve
+	float m_time;		// Max time that should take to move
+	bool m_stopAtEnd;   // When arrive at location should stop?
 
 	float m_accelerationTime; // % of the time it will accelerate
 	float m_decelerationTime; // % of the time it will decelerate
